third_party/WebKit/LayoutTests/fast/canvas/image-pattern-rotate-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS imgdata[0] is 0
PASS imgdata[1] is 255
PASS imgdata[2] is 0
PASS imgdata[0] is 0
PASS imgdata[1] is 255
PASS imgdata[2] is 0
PASS imgdata[0] is 0
PASS imgdata[1] is 255
PASS imgdata[2] is 0
PASS imgdata[0] is 0
PASS imgdata[1] is 255
PASS imgdata[2] is 0
PASS imgdata[0] is 0
PASS imgdata[1] is 255
PASS imgdata[2] is 0
PASS imgdata[0] is 0
PASS imgdata[1] is 255
PASS imgdata[2] is 0
PASS imgdata[0] is 0
PASS imgdata[1] is 255
PASS imgdata[2] is 0
PASS imgdata[0] is 0
PASS imgdata[1] is 255
PASS imgdata[2] is 0
PASS imgdata[0] is 0
PASS imgdata[1] is 255
PASS imgdata[2] is 0
PASS imgdata[0] is 0
PASS imgdata[1] is 255
PASS imgdata[2] is 0
PASS imgdata[0] is 0
PASS imgdata[1] is 255
PASS imgdata[2] is 0
PASS imgdata[0] is 0
PASS imgdata[1] is 255
PASS imgdata[2] is 0
PASS imgdata[0] is 0
PASS imgdata[1] is 255
PASS imgdata[2] is 0
PASS imgdata[0] is 0
PASS imgdata[1] is 255
PASS imgdata[2] is 0
PASS imgdata[0] is 0
PASS imgdata[1] is 255
PASS imgdata[2] is 0
PASS imgdata[0] is 0
PASS imgdata[1] is 255
PASS imgdata[2] is 0
Visual result below is for debugging purposes only (test uses text baseline only). You should see sixteen canvases with a green dot (rotated square) in the center.
